# Add optimized simultaneous ECDF bands