--- a/arviz/tests/base_tests/test_stats_ecdf_utils.py
+++ b/arviz/tests/base_tests/test_stats_ecdf_utils.py
@@ -11,7 +11,7 @@ from ...stats.ecdf_utils import (
 )
 
 try:
-    import numba
+    import numba   # pylint: disable=unused-import
 
     numba_options = [True, False]
 except ImportError: