@sethaxen 2024-08-22
> I'll need to think more about whether this would work for arbitrary ECDFs or only for cases where the CDF at eval points is a uniform grid between 0 and 1 (as is the case for uniform eval points for uniform distribution).

Yeah, that's the only context where this makes sense, which is the same context as PIT plots. Ergo, I'll leave interpolation for a separate PR, and in this PR I'll use a benchmark to heuristically identify when we should switch from optimized bands to simulated as the default in `plot_ecdf`.